add Wifi Screen-Off Power Management from franco.Kernel
This gives you the option to choose between PM_FAST and PM_MAX when your screen is off.
Normally the kernel uses PM_FAST only if the screen is on. But when the screen turns off, it switches to PM_MAX, which slows the wifi down to save battery.
But this can affect streaming applications (i.e. music). Which will result in lower quality audio or disconnects.

echo X > /sys/module/bcmdhd/parameters/wifi_pm
X can be 0 or 1. 0 (default) sets it to PM_MAX, and 1 sets it to PM_FAST.